fix(oracle): bound the clangd marker search and stop re-reading its database (#536)
Two costs on the path taken for every spawn attempt, while the maintenance pass
holds the repository write lock.

The marker walk follows directory symlinks on purpose — that is what makes
`build -> cmake-build-debug` discoverable — and the depth bound it relied on
caps recursion depth, not the number of paths explored. One self-referential
link costs a linear walk; two pointing at the same ancestor branch at every
level, so the walk is exponential in the bound. Replaying that traversal against
a tree with two root-level self-links did not finish in sixty seconds, which in
the watcher means a wedged pass holding the write lock. Only entries whose type
says symlink are canonicalized now — that check is answered from the readdir
result, so ordinary directories pay nothing — and a link whose target the search
already entered is not followed again. The depth bound stays as a backstop for a
merely deep tree.

Deciding whether a compilation database is usable opens it and parses its first
entry, and the per-file discovery path asked once per worklist path per pass.
Every source file in a directory shares the same nearest database, so a
two-hundred-file worklist re-read and re-parsed the same file two hundred times.
Verdicts are memoized on the project layout, seeded from the scan that produced
it, and die with it — so `LAYOUT_MAX_AGE` remains the only staleness window.

The prerequisite hint regressed when the TypeScript and clangd gates were
merged: both backends ask the same question, but the shared wording dropped the
symptom each server actually exhibits. The gate stays shared and the hint is per
backend again, so neither reads as advice for the other.

The registry also moves from `backend.rs` to a `backend/` directory — entries,
the project-layout search, the document searches, and the tests, with `mod.rs`
as the index. It lands here rather than separately because it relocates the code
this commit changes; the behavioural changes are the three above plus the
manifest comment, and everything else is a move.
