+++ crates/rag-rat-oracle/src/manifest.rs
@@ -138,8 +138,9 @@ impl ToolManifest {
             // here because it is load-bearing: it is what resolves a call across translation
             // units, and with it off clangd answers with the header declaration instead. It also
             // makes clangd persist an index into `$CHECKOUT/.cache/clangd/` — no flag or
-            // environment variable relocates that, so the write is accepted (and `.cache` is
-            // floored out of indexing so it cannot feed the watcher back into itself).
+            // environment variable relocates that, so the write is accepted, and that one path
+            // (`.cache/clangd`, not `.cache` as a whole) is floored out of the discovery walk so a
+            // large, entirely machine-written index tree is never indexed as first-party code.
             OracleTool::ClangdLsp => ToolManifest {
                 tool,
                 program: "clangd",
@@ -332,9 +333,11 @@ impl ToolManifest {
             // from. Without one a call resolves only to its header declaration, and clangd emits
             // no project-load progress at all — so the backend could never report ready.
             // Both progress-signalled live backends gate on the SAME question — "is there a
-            // project here whose load this server would report?" — so they share the check, and
-            // each names its own marker in the hint. `checkout_can_signal_readiness` is the same
-            // search the warm-up uses, so the gate and the warm-up cannot disagree.
+            // project here whose load this server would report?" — so they share the check.
+            // The HINT is per backend: the shared sentence states the gate, and the marker's own
+            // detail names the measured symptom and the command that fixes it, which differ per
+            // server. `checkout_can_signal_readiness` is the same search the warm-up uses, so the
+            // gate and the warm-up cannot disagree.
             OracleTool::ClangdLsp | OracleTool::TsLsp => {
                 let backend = super::backend::LiveBackend::for_tool(self.tool)?;
                 let resolved;
@@ -345,15 +348,22 @@ impl ToolManifest {
                         &resolved
                     },
                 };
+                // A progress-signalled backend that declared no marker could never signal either,
+                // so it still blocks — with the generic wording, since there is no file to name.
+                let (marker, detail) =
+                    backend.project_marker.map_or(("project", "Add one to enable it."), |marker| {
+                        (marker.file, marker.hint_detail)
+                    });
                 (!backend.checkout_can_signal_readiness(root, layout)).then(|| {
                     format!(
                         "the live {} oracle found no {} project under {} — {} only reports \
                          project-load progress for a real project, and that signal is what tells \
-                         the oracle its answers are trustworthy. Add one to enable it.",
-                        self.languages.join("/"),
-                        backend.project_marker.map_or("project", |marker| marker.file),
+                         the oracle its answers are trustworthy. {}",
+                        backend.display_name,
+                        marker,
                         root.display(),
                         self.program,
+                        detail,
                     )
                 })
             },
@@ -582,7 +592,15 @@ mod tests {
         // batch scip-clang backend requires, for its own reasons.
         let dir = rag_rat_base::test_scratch::ScratchDir::new("clangd-lsp-prereq");
         let blocked = manifest.prerequisite_blocked(&dir).expect("no compdb ⇒ Blocked");
+        // The two progress-signalled backends share the gate but not the hint: this one names
+        // clangd's marker, clangd's measured symptom, and the command that produces a database.
+        // Nothing here may read as the TypeScript backend's advice.
+        assert!(blocked.contains("the live C/C++ oracle"), "{blocked}");
         assert!(blocked.contains("compile_commands.json"), "{blocked}");
+        assert!(blocked.contains("header declaration"), "{blocked}");
+        assert!(blocked.contains("bear -- make"), "{blocked}");
+        assert!(!blocked.contains("tsconfig.json"), "{blocked}");
+        assert!(!blocked.contains("import statement"), "{blocked}");
         std::fs::create_dir_all(dir.join("src")).unwrap();
         std::fs::write(dir.join("src/main.c"), "int main(void) { return 0; }\n").unwrap();
         // A syntactically valid but EMPTY database describes no project: measured, clangd emits
@@ -611,7 +629,14 @@ mod tests {
         let manifest = ToolManifest::for_tool(OracleTool::TsLsp);
         let dir = rag_rat_base::test_scratch::ScratchDir::new("ts-lsp-prereq");
         let blocked = manifest.prerequisite_blocked(&dir).expect("no project ⇒ Blocked");
+        // Its own marker, its own measured symptom, its own nudge — and none of clangd's, which
+        // shares the gate and would otherwise be indistinguishable here.
+        assert!(blocked.contains("the live TypeScript oracle"), "{blocked}");
         assert!(blocked.contains("tsconfig.json"), "{blocked}");
+        assert!(blocked.contains("import statement instead of the definition"), "{blocked}");
+        assert!(blocked.contains("most TypeScript projects ship one"), "{blocked}");
+        assert!(!blocked.contains("compile_commands.json"), "{blocked}");
+        assert!(!blocked.contains("header declaration"), "{blocked}");
         // A config need not sit at the ROOT, and it must have a file to open: the progress cycle
         // reports a PROJECT load, so an empty project is nothing to warm on.
         std::fs::create_dir_all(dir.join("packages/app")).unwrap();
